feat: support unlimited k-mer counts and add sequential processing API
- Change all k-mer counts from i32 to u64 across the entire API,
  eliminating the ~2.1 billion count ceiling
- Add count_kmers_sequential() for memory-constrained environments
  that processes records one-at-a-time without batching
- Update streaming module docs with API comparison table

+++ examples/streaming_large_file.rs
@@ -31,7 +31,7 @@ fn main() {
 
     let path = &args[1];
     let k: usize = args.get(2).and_then(|s| s.parse().ok()).unwrap_or(21);
-    let min_count: i32 = args.get(3).and_then(|s| s.parse().ok()).unwrap_or(1);
+    let min_count: u64 = args.get(3).and_then(|s| s.parse().ok()).unwrap_or(1);
 
     eprintln!("Counting {k}-mers in {path} (streaming mode)...");
     eprintln!("Minimum count threshold: {min_count}");
